refactor(overlay): split priming + screens out of overlayservice.cpp (phase 5.2)
overlayservice.cpp went over the project's <800-line cap during the
Phase 2-4 work. Split two coherent chunks out:

  - overlayservice/priming.cpp (182 lines) holds
    primeSurfaceRenderPipeline + cancelSurfacePrime — the RHI/Vulkan
    pipeline-warmup helpers that wire fire-once frameSwapped-driven
    hide on freshly-created surfaces. Content-agnostic and grouped
    with itself.
  - overlayservice/screens.cpp (200 lines) holds setupForScreen /
    removeScreen / assertWindowOnScreen / handleScreenAdded /
    destroyAllWindowsForPhysicalScreen / handleScreenRemoved — the
    screen-lifecycle handlers.

overlayservice.cpp shrinks from 954 to 785 lines.

Build clean. 178/178 tests pass.
