+++ src/daemon/overlayservice.cpp
@@ -103,164 +103,9 @@ QQuickItem* OverlayService::PerScreenOverlayState::mainOverlaySlot() const
 // overlayservice/animation_config.cpp to keep this translation unit
 // under the project's <800-line guideline.
 
-void OverlayService::primeSurfaceRenderPipeline(PhosphorLayer::Surface* surface)
-{
-    if (!surface) {
-        return;
-    }
-    // contains-check covers BOTH lifecycle stages of a single prime
-    // (pending warm + window-armed): once a surface is in the set, no
-    // path adds another stateChanged or frameSwapped lambda to it.
-    // Without this gate, an external double-call to
-    // primeSurfaceRenderPipeline (e.g. show path that races a screen
-    // reconfigure) would arm a second frameSwapped connection — one
-    // would fire and hide the surface mid-content.
-    if (m_primingSurfaces.contains(surface)) {
-        return;
-    }
-
-    // Single destroyed-cleanup per surface (per OverlayService
-    // instance), tracked in m_primingDestroyedConnections. Replaces
-    // the earlier `pz_primingDestroyedConnected` dynamic-property
-    // gate which leaked across service instances — a fresh service
-    // re-encountering the same Surface* would skip wiring its own
-    // cleanup. The slot's static_cast on `dying` is safe because the
-    // resulting pointer is only used as a hash-map key (compare-by-
-    // address); ~QObject has already run by the time destroyed fires.
-    if (!m_primingDestroyedConnections.contains(surface)) {
-        QMetaObject::Connection destroyedConn = connect(surface, &QObject::destroyed, this, [this](QObject* dying) {
-            auto* surf = static_cast<PhosphorLayer::Surface*>(dying);
-            m_primingSurfaces.remove(surf);
-            m_primingFrameConnections.remove(surf);
-            m_primingDestroyedConnections.remove(surf);
-        });
-        m_primingDestroyedConnections.insert(surface, destroyedConn);
-    }
-
-    auto* window = surface->window();
-    if (!window) {
-        // Surface hasn't materialised a QQuickWindow yet — Surface::warmUp
-        // is asynchronous for content that compiles off the main thread.
-        // Defer until warm completes (stateChanged Warming → Hidden).
-        // Disconnect on the FIRST Hidden — even if the window is somehow
-        // still null we drop the connection rather than letting it stay
-        // armed forever and re-fire on every later state change. The
-        // recursive call lands in the window-non-null branch which adds
-        // to m_primingSurfaces and installs the frameSwapped path.
-        //
-        // Insert into m_primingSurfaces NOW (warm-pending sentinel) so
-        // an external second call to primeSurfaceRenderPipeline before
-        // the first warm completes hits the contains() guard above and
-        // bails — without this, the second call would queue a SECOND
-        // stateChanged lambda whose recursive call lands in the window-
-        // path's contains() bail at line `m_primingSurfaces.contains
-        // (surface) → return` after the first one already inserted +
-        // armed, leaking the second stateChanged slot for the rest of
-        // the surface's lifetime.
-        //
-        // Disconnects on Hidden OR Failed: a surface stuck in Failed
-        // never reaches Hidden, so without the Failed branch the
-        // sentinel sits in m_primingSurfaces indefinitely, blocking
-        // any future re-prime even after a recovery path puts the
-        // surface back into a usable state.
-        m_primingSurfaces.insert(surface);
-        QPointer<PhosphorLayer::Surface> guard(surface);
-        auto warmConn = std::make_shared<QMetaObject::Connection>();
-        *warmConn = connect(surface, &PhosphorLayer::Surface::stateChanged, this,
-                            [this, guard, warmConn](PhosphorLayer::Surface::State newState) {
-                                if (newState != PhosphorLayer::Surface::State::Hidden
-                                    && newState != PhosphorLayer::Surface::State::Failed) {
-                                    return;
-                                }
-                                QObject::disconnect(*warmConn);
-                                if (!guard) {
-                                    return;
-                                }
-                                // Drop the warm-pending sentinel BEFORE the
-                                // recursive call so the window-path's
-                                // contains() guard re-evaluates to false
-                                // and proceeds to insert + arm the
-                                // frameSwapped handler.
-                                //
-                                // CRITICAL: gate the recursive prime on
-                                // `m_primingSurfaces.remove(...)` returning
-                                // true. If a user-show called
-                                // cancelSurfacePrime during the warm-pending
-                                // window, the surface was already removed
-                                // from the set, `remove()` returns false,
-                                // and we MUST NOT recurse — recursion would
-                                // re-arm a fresh prime cycle whose
-                                // frameSwapped-driven hide() races the
-                                // user's just-shown content off the screen.
-                                // Without this guard, cancelSurfacePrime is
-                                // a silent no-op while the warm hasn't
-                                // completed.
-                                const bool stillPriming = m_primingSurfaces.remove(guard.data());
-                                if (stillPriming && newState == PhosphorLayer::Surface::State::Hidden
-                                    && guard->window() != nullptr) {
-                                    primeSurfaceRenderPipeline(guard.data());
-                                }
-                            });
-        return;
-    }
-    m_primingSurfaces.insert(surface);
-
-    // Hide on the first frameSwapped after surface->show(). By that
-    // point the wl_surface is mapped, the Vulkan swapchain has at
-    // least one image, and the QML scene-graph (including any
-    // QSGLayer that the shader path will later use) has rendered
-    // at least one frame.
-    //
-    // The connection is tracked in m_primingFrameConnections so
-    // cancelSurfacePrime can disconnect it explicitly — without
-    // tracking, the connection survives until next paint and we
-    // accumulate one stale slot per prime cycle for the surface's
-    // lifetime under rapid show/hide.
-    QPointer<PhosphorLayer::Surface> guard(surface);
-    QMetaObject::Connection frameConn = connect(window, &QQuickWindow::frameSwapped, this, [this, guard]() {
-        if (!guard) {
-            // Surface died after the connection was armed but before
-            // first frameSwapped — the destroyed-signal lambda in
-            // m_primingDestroyedConnections has already cleaned the
-            // map entry, and Qt's sender-destruction auto-disconnect
-            // (window dies with surface) will retire this lambda
-            // shortly. Nothing to do here.
-            return;
-        }
-        const auto connIt = m_primingFrameConnections.find(guard.data());
-        if (connIt != m_primingFrameConnections.end()) {
-            QObject::disconnect(connIt.value());
-            m_primingFrameConnections.erase(connIt);
-        }
-        // Only hide if the user hasn't already taken over the surface
-        // (cancelSurfacePrime would have removed us from the set).
-        if (m_primingSurfaces.remove(guard.data())) {
-            guard->hide();
-        }
-    });
-    m_primingFrameConnections.insert(surface, frameConn);
-    surface->show();
-}
-
-void OverlayService::cancelSurfacePrime(PhosphorLayer::Surface* surface)
-{
-    // Idempotent — called from every user show path so a non-priming
-    // surface short-circuits cheaply. Disconnect the frameSwapped
-    // lambda EXPLICITLY (tracked in m_primingFrameConnections) so the
-    // queued hide-on-first-paint never fires after a user-show. The
-    // m_primingSurfaces.remove() is the secondary guard the lambda
-    // would also check, but explicit disconnection is the safer
-    // primary contract — any future event-loop pump between cancel
-    // and the user's surface->show() is now harmless. Surfaces that
-    // get torn down outside of cancelSurfacePrime are cleaned via the
-    // destroyed signal connection in m_primingDestroyedConnections.
-    m_primingSurfaces.remove(surface);
-    const auto connIt = m_primingFrameConnections.find(surface);
-    if (connIt != m_primingFrameConnections.end()) {
-        QObject::disconnect(connIt.value());
-        m_primingFrameConnections.erase(connIt);
-    }
-}
+// primeSurfaceRenderPipeline + cancelSurfacePrime are extracted to
+// overlayservice/priming.cpp to keep this translation unit under the
+// project's <800-line guideline.
 
 OverlayService::OverlayService(Phosphor::Screens::ScreenManager* screenManager, ShaderRegistry* shaderRegistry,
                                PhosphorAnimation::PhosphorProfileRegistry* profileRegistry, QObject* parent)
@@ -832,179 +677,10 @@ void OverlayService::setCurrentActivity(const QString& activityId)
     }
 }
 
-void OverlayService::setupForScreen(QScreen* screen)
-{
-    // Set up overlay windows for all effective screens on this physical screen
-    auto* mgr = m_screenManager;
-    const QString physId = Phosphor::Screens::ScreenIdentity::identifierFor(screen);
-    if (mgr && mgr->hasVirtualScreens(physId)) {
-        for (const QString& vsId : mgr->virtualScreenIdsFor(physId)) {
-            if (!m_screenStates.contains(vsId) || !m_screenStates[vsId].overlayPhysScreen) {
-                QRect vsGeom = mgr->screenGeometry(vsId);
-                if (!vsGeom.isValid()) {
-                    qCWarning(lcOverlay) << "setupForScreen: invalid geometry for virtual screen" << vsId
-                                         << ", skipping overlay creation";
-                    continue;
-                }
-                createOverlayWindow(vsId, screen, vsGeom);
-            }
-        }
-    } else {
-        if (!m_screenStates.contains(physId) || !m_screenStates[physId].overlayPhysScreen) {
-            createOverlayWindow(screen);
-        }
-    }
-}
-
-void OverlayService::removeScreen(QScreen* screen)
-{
-    destroyOverlayWindow(screen);
-}
-
-void OverlayService::assertWindowOnScreen(QWindow* window, QScreen* screen, const QRect& geometry)
-{
-    if (!window || !screen) {
-        return;
-    }
-    if (window->screen() != screen) {
-        window->setScreen(screen);
-    }
-    // For virtual screens (geometry differs from physical), positioning is handled by
-    // LayerShellQt margins. Calling setGeometry with absolute coordinates would override
-    // those margins, causing double-positioning. Only set geometry for physical screens.
-    const QRect targetGeom = geometry.isValid() ? geometry : screen->geometry();
-    if (targetGeom == screen->geometry()) {
-        window->setGeometry(targetGeom);
-    }
-    // Virtual screens: size is set by the caller; position is set by LayerShellQt margins.
-}
-
-void OverlayService::handleScreenAdded(QScreen* screen)
-{
-    if (!screen) {
-        return;
-    }
-
-    // Always attempt to recreate overlay windows for reconnected screens,
-    // even when m_visible is false. This handles monitor power-cycle recovery:
-    // screenRemoved clears m_notificationCreationFailed sentinels, so when the
-    // screen comes back we must retry shell creation. If m_visible is true the
-    // existing path also shows the overlay; if false we just ensure the
-    // windows exist so the next showAtPosition call finds them ready.
-    const bool wasVisible = m_visible;
-    if (!wasVisible) {
-        m_visible = true; // transient — initializeOverlay may set it false again
-    }
-
-    const QString physScreenId = Phosphor::Screens::ScreenIdentity::identifierFor(screen);
-
-    auto* mgr = m_screenManager;
-    if (mgr && mgr->hasVirtualScreens(physScreenId)) {
-        // Create overlays for each virtual screen on this physical screen
-        for (const QString& vsId : mgr->virtualScreenIdsFor(physScreenId)) {
-            if (isContextDisabled(m_settings, PhosphorZones::AssignmentEntry::Snapping, vsId, m_currentVirtualDesktop,
-                                  m_currentActivity)) {
-                continue;
-            }
-            QRect vsGeom = mgr->screenGeometry(vsId);
-            if (vsGeom.isValid()) {
-                createOverlayWindow(vsId, screen, vsGeom);
-                updateOverlayWindow(vsId, screen);
-                const auto& vsState = m_screenStates.value(vsId);
-                if (vsState.overlayPhysScreen) {
-                    if (auto* window = vsState.shell->shellWindow) {
-                        assertWindowOnScreen(window, screen, vsGeom);
-                        if (vsState.shell->shellSurface && !vsState.shell->shellSurface->isLogicallyShown()) {
-                            vsState.shell->shellSurface->show();
-                        }
-                    }
-                }
-            }
-        }
-    } else {
-        createOverlayWindow(screen);
-        updateOverlayWindow(screen);
-        const auto& pState = m_screenStates.value(physScreenId);
-        if (pState.overlayPhysScreen) {
-            if (auto* window = pState.shell->shellWindow) {
-                assertWindowOnScreen(window, screen);
-                if (pState.shell->shellSurface && !pState.shell->shellSurface->isLogicallyShown()) {
-                    pState.shell->shellSurface->show();
-                }
-            }
-        }
-    }
-
-    // Restore m_visible if the caller wasn't already in a visible state and
-    // no overlay windows were actually created (transport unavailable, etc.).
-    if (!wasVisible && !m_visible) {
-        qCDebug(lcOverlay) << "handleScreenAdded: no overlay windows created for" << physScreenId
-                           << "— transport may be unavailable";
-    }
-}
-
-void OverlayService::destroyAllWindowsForPhysicalScreen(QScreen* screen)
-{
-    // Remove all windows associated with this physical screen
-    // (includes any virtual screens on this physical screen)
-    const QStringList screenIds = m_screenStates.keys();
-    for (const QString& id : screenIds) {
-        const auto& state = m_screenStates[id];
-        if (state.overlayPhysScreen == screen || state.zoneSelectorPhysScreen == screen
-            || state.shell->physScreen == screen) {
-            destroyOverlayWindow(id);
-            destroyZoneSelectorWindow(id);
-            destroyPassiveShell(id);
-            // Drop the empty state entry once the shell surface is
-            // gone — matches cleanupVirtualScreenStates semantics so
-            // screen hot-plug cycles don't slowly accumulate dead keys.
-            auto& s = m_screenStates[id];
-            if (!s.shell->shellSurface) {
-                m_screenStates.remove(id);
-            }
-        }
-    }
-
-    // Snap-assist + layout picker post-shell-migration are Item slots
-    // inside the per-screen passive shell — destroying the shell
-    // (above, via destroyPassiveShell) tears the slots down with
-    // it. No separate cleanup needed.
-
-    // Drop notification-window "creation failed" sentinels for screen ids
-    // rooted on this physical screen. Without this, if the same physical
-    // monitor is reconnected (hot-plug cycle) it inherits the stale flag
-    // and we silently refuse to recreate the OSD. Matching is prefix-based
-    // because virtual-screen ids embed the physical id as the prefix.
-    const QString physId = Phosphor::Screens::ScreenIdentity::identifierFor(screen);
-    if (!physId.isEmpty()) {
-        const QString vsPrefix = physId + PhosphorIdentity::VirtualScreenId::Separator;
-        for (const QString& flagged : m_shellHost->failureScreenIds()) {
-            if (flagged == physId || flagged.startsWith(vsPrefix)) {
-                m_shellHost->clearFailure(flagged);
-            }
-        }
-    }
-
-    // Drop the dedup sentinel for this physical screen so a hot-plug cycle
-    // doesn't suppress the first navigation OSD on the reconnected monitor
-    // when it lands inside the implicit 200 ms timeout window. The dedup is
-    // keyed on the screenId at time-of-fire, and a removed-then-readded
-    // monitor reuses the same id — without this clear, a navigation action
-    // on the readded screen within 200 ms of the last action on the
-    // pre-removal incarnation gets silently swallowed.
-    if (!physId.isEmpty()
-        && (m_lastNavigationScreenId == physId
-            || m_lastNavigationScreenId.startsWith(physId + PhosphorIdentity::VirtualScreenId::Separator))) {
-        m_lastNavigationActionKey.clear();
-        m_lastNavigationScreenId.clear();
-        m_lastNavigationTime.invalidate();
-    }
-}
-
-void OverlayService::handleScreenRemoved(QScreen* screen)
-{
-    destroyAllWindowsForPhysicalScreen(screen);
-}
+// Screen-management methods (setupForScreen / removeScreen /
+// assertWindowOnScreen / handleScreenAdded / destroyAllWindowsForPhysicalScreen
+// / handleScreenRemoved) are extracted to overlayservice/screens.cpp to keep
+// this translation unit under the project's <800-line guideline.
 
 OverlayService::LayoutIncludeFlags OverlayService::resolvePerScreenLayoutInclude(const QString& screenId) const
 {
